Move insert_in_repo() to metadata.py.
insert_in_repo() is only used by bodhi.server.metadata, but it had
been in bodhi.server.util. This added an unnecessary dependency on
createrepo_c to the web server. By moving it into metadata.py, this
dependency is avoided.

fixes #2565
